move change detection script to application folder
